refactor: rename form-settings to form-state
