+++ packages/form-builder/src/components/canvas/DesignPreview.tsx
@@ -4,7 +4,7 @@ import {useEffect, useState} from 'react';
 import Storage from '@givewp/form-builder/common/storage';
 
 import IframeResizer from 'iframe-resizer-react';
-import {useFormState} from '../../stores/form-settings';
+import {useFormState} from '../../stores/form-state';
 
 const DesignPreview = () => {
     const {
