+++ packages/form-builder/src/settings/donation-goal/index.js
@@ -1,4 +1,4 @@
-import {setFormSettings, useFormState, useFormStateDispatch} from '@givewp/form-builder/stores/form-settings';
+import {setFormSettings, useFormState, useFormStateDispatch} from '@givewp/form-builder/stores/form-state';
 import {__} from '@wordpress/i18n';
 import {
     __experimentalNumberControl as NumberControl,
