+++ packages/form-builder/src/settings/form-title/index.js
@@ -1,6 +1,6 @@
 import {PanelBody, PanelRow, TextControl} from '@wordpress/components';
 import {__} from '@wordpress/i18n';
-import {setFormSettings, useFormState, useFormStateDispatch} from '@givewp/form-builder/stores/form-settings';
+import {setFormSettings, useFormState, useFormStateDispatch} from '@givewp/form-builder/stores/form-state';
 
 const FormTitleSettings = () => {
     const {
