+++ packages/form-builder/src/settings/offline-donation/index.js
@@ -1,4 +1,4 @@
-import {setFormSettings, useFormState, useFormStateDispatch} from '@givewp/form-builder/stores/form-settings';
+import {setFormSettings, useFormState, useFormStateDispatch} from '@givewp/form-builder/stores/form-state';
 import {PanelBody, PanelRow, ToggleControl} from '@wordpress/components';
 import {__} from '@wordpress/i18n';
 import DonationInstructions from './donation-instructions';
